Add feature gates to the `sap-aria` crate
`sap-aria` now has the following features:
- name - accessibility name fn
- property - AriaProperty enum and dependents
- role - AriaRole enum and dependents
- state - AriaState enum and dependents

`sap-aria` defaults to using all the above features
as they are all required in `sap`.
